Create volatileacidityvsquality.py
On running the above code, it is clearly evident from the graph that it is an inverse relationship as the slope is negative. Also from the correlation value r for this case comes out to be -0.39 which shows that it a negative and weak relation impact of volatile acidity and quality of Wine.